some refactoring and fix infinite loading
