+++ src/components/statusBarItem.ts
@@ -1,0 +1,15 @@
+import * as vscode from 'vscode'
+
+const STATUS_BAR_ITEM: string = 'devcycle-featureflags'
+export const statusBarItem = vscode.window.createStatusBarItem(STATUS_BAR_ITEM)
+statusBarItem.name = 'DevCycle Status'
+
+export function showBusyMessage(message: string) {
+    statusBarItem.text = `$(loading~spin) ${message}...`
+    statusBarItem.tooltip = `${message}...`
+    statusBarItem.show()
+  }
+  
+export function hideBusyMessage() {
+    statusBarItem.hide()
+}
